feat: move bool_trans from storage to spec
