+++ custom_components/xiaomi_home/miot/miot_storage.py
@@ -719,89 +719,6 @@ class MIoTCert:
         return binascii.hexlify(sha1_hash.finalize()).decode('utf-8')
 
 
-class SpecBoolTranslation:
-    """
-    Boolean value translation.
-    """
-    BOOL_TRANS_FILE = 'specs/bool_trans.json'
-    _main_loop: asyncio.AbstractEventLoop
-    _lang: str
-    _data: Optional[dict[str, dict]]
-    _data_default: dict[str, dict]
-
-    def __init__(
-        self, lang: str, loop: Optional[asyncio.AbstractEventLoop] = None
-    ) -> None:
-        self._main_loop = loop or asyncio.get_event_loop()
-        self._lang = lang
-        self._data = None
-
-    async def init_async(self) -> None:
-        if isinstance(self._data, dict):
-            return
-        data = None
-        self._data = {}
-        try:
-            data = await self._main_loop.run_in_executor(
-                None, load_json_file,
-                os.path.join(
-                    os.path.dirname(os.path.abspath(__file__)),
-                    self.BOOL_TRANS_FILE))
-        except Exception as err:  # pylint: disable=broad-exception-caught
-            _LOGGER.error('bool trans, load file error, %s', err)
-            return
-        # Check if the file is a valid JSON file
-        if (
-            not isinstance(data, dict)
-            or 'data' not in data
-            or not isinstance(data['data'], dict)
-            or 'translate' not in data
-            or not isinstance(data['translate'], dict)
-        ):
-            _LOGGER.error('bool trans, valid file')
-            return
-
-        if 'default' in data['translate']:
-            data_default = (
-                data['translate']['default'].get(self._lang, None)
-                or data['translate']['default'].get(
-                    DEFAULT_INTEGRATION_LANGUAGE, None))
-            if data_default:
-                self._data_default = [
-                    {'value': True, 'description': data_default['true']},
-                    {'value': False, 'description': data_default['false']}
-                ]
-
-        for urn, key in data['data'].items():
-            if key not in data['translate']:
-                _LOGGER.error('bool trans, unknown key, %s, %s', urn, key)
-                continue
-            trans_data = (
-                data['translate'][key].get(self._lang, None)
-                or data['translate'][key].get(
-                    DEFAULT_INTEGRATION_LANGUAGE, None))
-            if trans_data:
-                self._data[urn] = [
-                    {'value': True, 'description': trans_data['true']},
-                    {'value': False, 'description': trans_data['false']}
-                ]
-
-    async def deinit_async(self) -> None:
-        self._data = None
-        self._data_default = None
-
-    async def translate_async(self, urn: str) -> list[dict[bool, str]]:
-        """
-        MUST call init_async() before calling this method.
-        [
-            {'value': True, 'description': 'True'},
-            {'value': False, 'description': 'False'}
-        ]
-        """
-
-        return self._data.get(urn, self._data_default)
-
-
 class SpecFilter:
     """
     MIoT-Spec-V2 filter for entity conversion.
